requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Supabase
supabase==2.10.0

# Google Sheets
gspread==6.1.4
google-auth==2.37.0

# Grok AI (OpenAI compatible SDK)
openai==1.59.9

# PDF Generation
fpdf2==2.8.3
jinja2==3.1.5

# HTTP / Retry
httpx==0.27.2
tenacity==9.0.0

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
loguru==0.7.3
